Add the game cards and sort function
Add the parameters from the API trough the databinding in the view.
Creat a function that sort the game by a paramter selected in the
dropdown.
Create a function call the API request at the load of the home.
Create the function to search the games from the searchbar in the
navbar.